pyproject.toml
[tool.poetry]
name = "health-pipeline"
version = "1.0.0"
description = "Resilient health insurance data pipeline – AWS + dbt + PostgreSQL"
authors = ["Data Engineering Assessment"]
readme = "README.md"
python = "^3.11"

[tool.poetry.dependencies]
python          = "^3.11"
pandas          = "^2.2"
pyarrow         = "^16.0"
boto3           = "^1.34"
psycopg2-binary = "^2.9"
dbt-core        = "^1.8"
dbt-postgres    = "^1.8"
pydantic        = "^2.7"

[tool.poetry.group.dev.dependencies]
pytest          = "^8.2"
pytest-cov      = "^5.0"
black           = "^24.4"
ruff            = "^0.4"
moto            = {extras = ["s3"], version = "^5.0"}

[build-system]
requires      = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length    = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
select      = ["E", "F", "I", "N", "W"]
